refactor(pfdi): Split ALP0 tests to enforce one rule per test method
 - Updated all tests to follow single rule per test case methodology
 - Removed shared logic and consolidated test responsibilities
 - Cleaned up global memory handling and removed unused allocations
