+++ test_pool/pfdi/pfdi001.c
@@ -16,150 +16,49 @@
  **/
 
 #include "val/include/acs_val.h"
-#include "val/include/acs_memory.h"
 #include "val/include/val_interface.h"
 
 #define TEST_NUM   (ACS_PFDI_TEST_NUM_BASE + 1)
-#define TEST_RULE  "R0053, R0055, R0104"
-#define TEST_DESC  "Check PFDI Version in All PE's            "
+#define TEST_RULE  "R0053"
+#define TEST_DESC  "Check PFDI Version is returned            "
 
-extern int64_t g_pfdi_version;
-
-typedef struct{
+static void payload_version_check(void *arg)
+{
+  uint32_t index = val_pe_get_index_mpid(val_pe_get_mpid());
+  uint32_t major = 0, minor = 0;
   int64_t pfdi_version;
-} version_details;
-
-version_details *g_pfdi_version_details;
-
-void
-pfdi_version_check(void)
-{
-  uint32_t index = val_pe_get_index_mpid(val_pe_get_mpid());
-  version_details *pfdi_status_buffer;
-
-  pfdi_status_buffer = g_pfdi_version_details + index;
 
   /* Invoke PFDI Version function for current PE index */
-  pfdi_status_buffer->pfdi_version = val_pfdi_version();
-
-  val_data_cache_ops_by_va((addr_t)&pfdi_status_buffer->pfdi_version, CLEAN_AND_INVALIDATE);
-  val_set_status(index, RESULT_PASS(TEST_NUM, 1));
-
-  return;
-}
-
-static void payload(uint32_t num_pe)
-{
-  uint32_t index = val_pe_get_index_mpid(val_pe_get_mpid());
-  uint32_t timeout, i = 0, major = 0, minor = 0;
-  version_details *pfdi_buffer;
-  int64_t version = 0;
-
-  /* Invoke PFDI Version function for current PE index */
-  g_pfdi_version = val_pfdi_version();
-  if (g_pfdi_version >> 31) {
+  pfdi_version = val_pfdi_version();
+  if (pfdi_version >> 31) {
     val_print(ACS_PRINT_ERR,
-                "\n       PFDI get version failed err = %d", g_pfdi_version);
+                "\n       PFDI get version failed err = %ld", pfdi_version);
     val_set_status(index, RESULT_FAIL(TEST_NUM, 1));
     return;
   }
 
   /* Return status Bits[63:31] must be zero */
   if (val_pfdi_reserved_bits_check_is_zero(
-             VAL_EXTRACT_BITS(g_pfdi_version, 31, 63)) != ACS_STATUS_PASS) {
+             VAL_EXTRACT_BITS(pfdi_version, 31, 63)) != ACS_STATUS_PASS) {
     val_print(ACS_PRINT_ERR, "\n       Failed on PE = %d", index);
     val_set_status(index, RESULT_FAIL(TEST_NUM, 2));
   }
 
-  major = PFDI_VERSION_GET_MAJOR(g_pfdi_version);
+  major = PFDI_VERSION_GET_MAJOR(pfdi_version);
   if (major != PFDI_MAJOR_VERSION) {
     val_print(ACS_PRINT_ERR, "\n       Major Version not as expected, Current version = %d", major);
     val_set_status(index, RESULT_FAIL(TEST_NUM, 3));
     return;
   }
 
-  minor = PFDI_VERSION_GET_MINOR(g_pfdi_version);
+  minor = PFDI_VERSION_GET_MINOR(pfdi_version);
   if (minor != PFDI_MINOR_VERSION) {
     val_print(ACS_PRINT_ERR, "\n       Minor Version not as expected, Current version = %d", minor);
     val_set_status(index, RESULT_FAIL(TEST_NUM, 4));
     return;
   }
 
-  /* Allocate memory to save all PFDI Versions or status for all PE's */
-  g_pfdi_version_details = (version_details *) val_memory_calloc(num_pe, sizeof(version_details));
-  if (g_pfdi_version_details == NULL) {
-    val_print(ACS_PRINT_ERR, "\n       Allocation for PFDI Version Details Failed \n", 0);
-    val_set_status(index, RESULT_FAIL(TEST_NUM, 5));
-    return;
-  }
-
-  for (i = 0; i < num_pe; i++) {
-    pfdi_buffer = g_pfdi_version_details + i;
-    val_data_cache_ops_by_va((addr_t)&pfdi_buffer->pfdi_version, CLEAN_AND_INVALIDATE);
-  }
-
-  /* Execute pfdi_version_check function in All PE's */
-  for (i = 0; i < num_pe; i++) {
-    if (i != index) {
-      timeout = TIMEOUT_LARGE;
-      val_execute_on_pe(i, pfdi_version_check, (uint64_t)g_pfdi_version_details);
-
-      while ((--timeout) && (IS_RESULT_PENDING(val_get_status(i))));
-
-      if (timeout == 0) {
-        val_print(ACS_PRINT_ERR, "\n       **Timed out** for PE index = %d", i);
-        val_set_status(i, RESULT_FAIL(TEST_NUM, 6));
-        return;
-      }
-    }
-  }
-  val_time_delay_ms(ONE_MILLISECOND);
-
-  for (i = 0; i < num_pe; i++) {
-    if (i != index) {
-      pfdi_buffer = g_pfdi_version_details + i;
-
-      val_data_cache_ops_by_va((addr_t)&pfdi_buffer->pfdi_version, CLEAN_AND_INVALIDATE);
-      version = pfdi_buffer->pfdi_version;
-      if (version >> 31) {
-        val_print(ACS_PRINT_ERR,
-                "\n       PFDI get version failed err = %d", version);
-        val_print(ACS_PRINT_ERR, "\n       Failed on PE = %d", i);
-        val_set_status(i, RESULT_FAIL(TEST_NUM, 7));
-        goto free_pfdi_details;
-      }
-
-      major = PFDI_VERSION_GET_MAJOR(version);
-      if (major != PFDI_MAJOR_VERSION) {
-        val_print(ACS_PRINT_ERR,
-                "\n       Major Version not as expected, Current version = 0x%llx", major);
-        val_print(ACS_PRINT_ERR, "\n       Failed on PE = %d", i);
-        val_set_status(i, RESULT_FAIL(TEST_NUM, 8));
-        goto free_pfdi_details;
-      }
-
-      minor = PFDI_VERSION_GET_MINOR(version);
-      if (minor != PFDI_MINOR_VERSION) {
-        val_print(ACS_PRINT_ERR,
-                "\n       Minor Version not as expected, Current version =%d", minor);
-        val_print(ACS_PRINT_ERR, "\n       Failed on PE = %d", i);
-        val_set_status(i, RESULT_FAIL(TEST_NUM, 9));
-        goto free_pfdi_details;
-      }
-
-      if (val_pfdi_reserved_bits_check_is_zero(
-               VAL_EXTRACT_BITS(pfdi_buffer->pfdi_version, 31, 63)) != ACS_STATUS_PASS) {
-        val_print(ACS_PRINT_ERR, "\n       Failed on PE = %d", i);
-        val_set_status(i, RESULT_FAIL(TEST_NUM, 10));
-        goto free_pfdi_details;
-      }
-    }
-  }
-
   val_set_status(index, RESULT_PASS(TEST_NUM, 1));
-
-free_pfdi_details:
-  val_memory_free((void *) g_pfdi_version_details);
 
   return;
 }
@@ -168,10 +67,11 @@ uint32_t pfdi001_entry(uint32_t num_pe)
 {
   uint32_t status = ACS_STATUS_FAIL;
 
+  num_pe = 1; /* This test will run on single PE */
   status = val_initialize_test(TEST_NUM, TEST_DESC, num_pe);
 
   if (status != ACS_STATUS_SKIP)
-    payload(num_pe);
+    val_run_test_configurable_payload(&num_pe, payload_version_check);
 
   status = val_check_for_error(TEST_NUM, num_pe, TEST_RULE);
   val_report_status(0, ACS_END(TEST_NUM), NULL);
