+++ test_pool/pfdi/pfdi003.c
@@ -20,105 +20,71 @@
 #include "val/include/acs_memory.h"
 
 #define TEST_NUM   (ACS_PFDI_TEST_NUM_BASE + 3)
-#define TEST_RULE  "R0066"
-#define TEST_DESC  "Check PE HW test mechanism info           "
+#define TEST_RULE  "R0102"
+#define TEST_DESC  "Check PFDI mandatory functions            "
 
 typedef struct{
-  int64_t pfdi_st_version;
-  int64_t pfdi_st_status;
-} st_version_details;
+  int64_t status[PFDI_FN_PFDI_MAX_NUM];
+} feature_details;
 
-st_version_details *g_pfdi_st_version_details;
+feature_details *g_pfdi_feature_details;
+
 
 void
-pfdi_st_version_check(void)
+pfdi_function_check(void)
 {
   uint32_t index = val_pe_get_index_mpid(val_pe_get_mpid());
-  st_version_details *pfdi_st_status_buffer;
-  int64_t feat;
+  uint32_t fn_status = 0, f_id;
+  feature_details *pfdi_buffer = g_pfdi_feature_details + index;
 
-  pfdi_st_status_buffer = g_pfdi_st_version_details + index;
-
-  /* Invoke PFDI Test ID  function for current PE index */
-  pfdi_st_status_buffer->pfdi_st_status = val_pfdi_pe_test_id(&feat);
-  pfdi_st_status_buffer->pfdi_st_version = feat;
-
-  val_data_cache_ops_by_va((addr_t)&pfdi_st_status_buffer->pfdi_st_version, CLEAN_AND_INVALIDATE);
-  val_data_cache_ops_by_va((addr_t)&pfdi_st_status_buffer->pfdi_st_status, CLEAN_AND_INVALIDATE);
+  for (f_id = PFDI_FN_PFDI_VERSION; f_id <= PFDI_FN_PFDI_FORCE_ERROR; f_id++) {
+    /* Invoke PFDI Feature function for current PE index */
+    pfdi_buffer->status[fn_status] = val_pfdi_features(f_id);
+    val_data_cache_ops_by_va((addr_t)&pfdi_buffer->status[fn_status], CLEAN_AND_INVALIDATE);
+    fn_status++;
+  }
 
   val_set_status(index, RESULT_PASS(TEST_NUM, 1));
   return;
 }
 
-static void payload(uint32_t num_pe)
+static void payload_functions_check(void *arg)
 {
   uint32_t index = val_pe_get_index_mpid(val_pe_get_mpid());
-  int64_t status, feat;
-  int64_t version;
-  int64_t temp_status;
-  uint32_t timeout, i = 0, major, minor, vendor_id;
-  st_version_details *st_pfdi_buffer;
+  uint32_t num_pe = *((uint32_t *)arg);
+  uint32_t f_id, fn_status = 0;
+  uint32_t timeout, i = 0, test_fail = 0;
+  feature_details *pfdi_buffer;
 
-  /* Invoke PFDI PE Test ID function for current PE index */
-  status = val_pfdi_pe_test_id(&feat);
-  if (status != PFDI_ACS_SUCCESS) {
-    val_print(ACS_PRINT_ERR,
-                "\n       PFDI PE Test ID failed err = %d", status);
+  /* Allocate memory to save all PFDI features status for all PE's */
+  g_pfdi_feature_details = (feature_details *) val_memory_calloc(num_pe, sizeof(feature_details));
+  if (g_pfdi_feature_details == NULL) {
+    val_print(ACS_PRINT_ERR, "\n       Allocation for PFDI Feature Details Failed \n", 0);
     val_set_status(index, RESULT_FAIL(TEST_NUM, 1));
     return;
   }
 
-  val_print(ACS_PRINT_TEST, "\n       PFDI Self Test version details for PE index = %d", index);
-  /* Return status Bits[63:32] must be zero */
-  if (val_pfdi_reserved_bits_check_is_zero(
-             VAL_EXTRACT_BITS(feat, 32, 63)) != ACS_STATUS_PASS) {
-    val_print(ACS_PRINT_ERR, "\n       Failed on PE = %d", index);
-    val_set_status(index, RESULT_FAIL(TEST_NUM, 2));
-  }
-
-  /* Return status Bits[23:20] must be zero */
-  if (val_pfdi_reserved_bits_check_is_zero(
-             VAL_EXTRACT_BITS(feat, 20, 23)) != ACS_STATUS_PASS) {
-    val_print(ACS_PRINT_ERR, "\n       Failed on PE = %d", index);
-    val_set_status(index, RESULT_FAIL(TEST_NUM, 3));
-  }
-
-  major = VAL_EXTRACT_BITS(feat, 8, 15);
-  val_print(ACS_PRINT_TEST, "\n       PFDI Self Test Major Version = %d", major);
-
-  minor = VAL_EXTRACT_BITS(feat, 0, 7);
-  val_print(ACS_PRINT_TEST, "\n       PFDI Self Test Minor Version = %d", minor);
-
-  vendor_id = VAL_EXTRACT_BITS(feat, 24, 31);
-  val_print(ACS_PRINT_TEST, "\n       PFDI Self Test Vendor ID     = %d", vendor_id);
-
-  /* Allocate memory to save all PFDI Self Test Versions for all PE's */
-  g_pfdi_st_version_details =
-            (st_version_details *) val_memory_calloc(num_pe, sizeof(st_version_details));
-  if (g_pfdi_st_version_details == NULL) {
-    val_print(ACS_PRINT_ERR,
-                "\n       Allocation for PFDI Self Test Version Details Failed", 0);
-    val_set_status(index, RESULT_FAIL(TEST_NUM, 4));
-    return;
-  }
-
   for (i = 0; i < num_pe; i++) {
-    st_pfdi_buffer = g_pfdi_st_version_details + i;
-    val_data_cache_ops_by_va((addr_t)&st_pfdi_buffer->pfdi_st_version, CLEAN_AND_INVALIDATE);
-    val_data_cache_ops_by_va((addr_t)&st_pfdi_buffer->pfdi_st_status, CLEAN_AND_INVALIDATE);
+    pfdi_buffer = g_pfdi_feature_details + i;
+    for (f_id = PFDI_FN_PFDI_VERSION; f_id <= PFDI_FN_PFDI_FORCE_ERROR; f_id++)
+      val_data_cache_ops_by_va((addr_t)&pfdi_buffer->status[fn_status++], CLEAN_AND_INVALIDATE);
   }
 
-  /* Execute pfdi_version_check function in All PE's */
+
+  /* Invoke PFDI Feature function for current PE index */
+  pfdi_function_check();
+
+  /* Execute pfdi_function_check function in All PE's */
   for (i = 0; i < num_pe; i++) {
     if (i != index) {
       timeout = TIMEOUT_LARGE;
-      val_execute_on_pe(i, pfdi_st_version_check, (uint64_t)g_pfdi_st_version_details);
+      val_execute_on_pe(i, pfdi_function_check, 0);
 
       while ((--timeout) && (IS_RESULT_PENDING(val_get_status(i))));
 
       if (timeout == 0) {
         val_print(ACS_PRINT_ERR, "\n       **Timed out** for PE index = %d", i);
-        val_set_status(i, RESULT_FAIL(TEST_NUM, 5));
+        val_set_status(i, RESULT_FAIL(TEST_NUM, 2));
         goto free_pfdi_details;
       }
     }
@@ -126,54 +92,33 @@ static void payload(uint32_t num_pe)
   val_time_delay_ms(ONE_MILLISECOND);
 
   for (i = 0; i < num_pe; i++) {
-    if (i != index) {
-      val_print(ACS_PRINT_DEBUG, "\n       PFDI Self Test version details for PE index = %d", i);
-      st_pfdi_buffer = g_pfdi_st_version_details + i;
+    pfdi_buffer = g_pfdi_feature_details + i;
+    test_fail = 0;
+    fn_status = 0;
 
-      val_data_cache_ops_by_va((addr_t)&st_pfdi_buffer->pfdi_st_status, CLEAN_AND_INVALIDATE);
-      val_data_cache_ops_by_va((addr_t)&st_pfdi_buffer->pfdi_st_version, CLEAN_AND_INVALIDATE);
+    for (f_id = PFDI_FN_PFDI_VERSION; f_id <= PFDI_FN_PFDI_FORCE_ERROR; f_id++)
+      val_data_cache_ops_by_va((addr_t)&pfdi_buffer->status[fn_status++], CLEAN_AND_INVALIDATE);
 
-      version = st_pfdi_buffer->pfdi_st_version;
-      temp_status  = st_pfdi_buffer->pfdi_st_status;
-
-      if (temp_status != PFDI_ACS_SUCCESS) {
+    fn_status = 0;
+    for (f_id = PFDI_FN_PFDI_VERSION; f_id <= PFDI_FN_PFDI_FORCE_ERROR; f_id++) {
+      if (pfdi_buffer->status[fn_status] != PFDI_ACS_SUCCESS) {
         val_print(ACS_PRINT_ERR,
-                "\n       PFDI PE Test ID failed err = %d", temp_status);
-        val_set_status(i, RESULT_FAIL(TEST_NUM, 6));
-        goto free_pfdi_details;
+                "\n       PFDI mandatory function 0x%x ", f_id);
+        val_print(ACS_PRINT_ERR, "failed on PE %d, ", i);
+        val_print(ACS_PRINT_ERR, "status = %d", pfdi_buffer->status[fn_status]);
+        test_fail++;
       }
+      fn_status++;
+    }
 
-      /* Return status Bits[63:32] must be zero */
-      if (val_pfdi_reserved_bits_check_is_zero(
-             VAL_EXTRACT_BITS(version, 32, 63)) != ACS_STATUS_PASS) {
-        val_print(ACS_PRINT_ERR, "\n       Failed on PE = %d", i);
-        val_set_status(i, RESULT_FAIL(TEST_NUM, 7));
-        goto free_pfdi_details;
-      }
-
-      /* Return status Bits[23:20] must be zero */
-      if (val_pfdi_reserved_bits_check_is_zero(
-             VAL_EXTRACT_BITS(version, 20, 23)) != ACS_STATUS_PASS) {
-        val_print(ACS_PRINT_ERR, "\n       Failed on PE = %d", i);
-        val_set_status(i, RESULT_FAIL(TEST_NUM, 8));
-        goto free_pfdi_details;
-      }
-
-      major = VAL_EXTRACT_BITS(version, 8, 15);
-      val_print(ACS_PRINT_DEBUG, "\n       PFDI Self Test Major Version = %d", major);
-
-      minor = VAL_EXTRACT_BITS(version, 0, 7);
-      val_print(ACS_PRINT_DEBUG, "\n       PFDI Self Test Minor Version = %d", minor);
-
-      vendor_id = VAL_EXTRACT_BITS(version, 24, 31);
-      val_print(ACS_PRINT_DEBUG, "\n       PFDI Self Test Vendor ID     = %d", vendor_id);
-   }
+    if (test_fail)
+      val_set_status(i, RESULT_FAIL(TEST_NUM, 3));
+    else
+      val_set_status(i, RESULT_PASS(TEST_NUM, 1));
   }
 
-  val_set_status(index, RESULT_PASS(TEST_NUM, 1));
-
 free_pfdi_details:
-  val_memory_free((void *) g_pfdi_st_version_details);
+  val_memory_free((void *) g_pfdi_feature_details);
 
   return;
 }
@@ -185,8 +130,9 @@ uint32_t pfdi003_entry(uint32_t num_pe)
   status = val_initialize_test(TEST_NUM, TEST_DESC, num_pe);
 
   if (status != ACS_STATUS_SKIP)
-    payload(num_pe);
+    val_run_test_configurable_payload(&num_pe, payload_functions_check);
 
+  /* get the result from all PE and check for failure */
   status = val_check_for_error(TEST_NUM, num_pe, TEST_RULE);
   val_report_status(0, ACS_END(TEST_NUM), NULL);
