+++ test_pool/pfdi/pfdi004.c
@@ -17,68 +17,85 @@
 
 #include "val/include/acs_val.h"
 #include "val/include/val_interface.h"
+#include "val/include/acs_memory.h"
 
 #define TEST_NUM   (ACS_PFDI_TEST_NUM_BASE + 4)
-#define TEST_RULE  "R0071"
-#define TEST_DESC  "Check num of Test Part supported          "
+#define TEST_RULE  "R0060"
+#define TEST_DESC  "Check PFDI Feature function support       "
 
-extern pfdi_pe_test_support_info *g_pfdi_pe_test_support_info;
+typedef struct{
+  int64_t status;
+  int64_t status_invalid;
+} pfdi_feature_check_details;
+
+pfdi_feature_check_details *g_pfdi_feature_check_details;
+
 
 void
-pfdi_test_part_count(void)
+check_feature()
 {
   uint32_t index = val_pe_get_index_mpid(val_pe_get_mpid());
-  int64_t  test_count;
-  pfdi_pe_test_support_info *test_buffer;
+  int64_t  status;
+  pfdi_feature_check_details  *status_buffer;
 
-  test_buffer = g_pfdi_pe_test_support_info + index;
+  status_buffer = g_pfdi_feature_check_details + index;
 
   /* Invoke PFDI Feature function for current PE index */
-  test_count = val_pfdi_pe_test_part_count();
-  if (test_count < PFDI_ACS_SUCCESS) {
-    val_set_status(index, RESULT_FAIL(TEST_NUM, 1));
-    return;
-  }
+  status = val_pfdi_features(PFDI_FN_PFDI_FEATURES);
 
-  test_buffer->test_count = test_count;
-  val_data_cache_ops_by_va((addr_t)&test_buffer->test_count, CLEAN_AND_INVALIDATE);
+  /*Save the return status*/
+  status_buffer->status = status;
+  val_data_cache_ops_by_va((addr_t)&status_buffer->status, CLEAN_AND_INVALIDATE);
+
+  /* Invoke PFDI Feature function with invalid id for current PE index */
+  status = val_pfdi_features(PFDI_FN_PFDI_INVALID);
+
+  /*Save the return status*/
+  status_buffer->status_invalid = status;
+  val_data_cache_ops_by_va((addr_t)&status_buffer->status_invalid, CLEAN_AND_INVALIDATE);
+
   val_set_status(index, RESULT_PASS(TEST_NUM, 1));
   return;
 }
 
-static void payload(uint32_t num_pe)
+static void payload_feature_check(void *arg)
 {
   uint32_t index = val_pe_get_index_mpid(val_pe_get_mpid());
-  uint32_t timeout, i = 0;
-  int64_t  test_count;
-  pfdi_pe_test_support_info *test_buffer;
+  uint32_t timeout, i = 0, run_fail = 0;
+  pfdi_feature_check_details *status_buffer;
+  uint32_t num_pe = *(uint32_t *)arg;
 
-  test_buffer = g_pfdi_pe_test_support_info + index;
-
-  /* Invoke PFDI Test Part count function for current PE index */
-  test_count = val_pfdi_pe_test_part_count();
-  if (test_count < PFDI_ACS_SUCCESS) {
+  /* Allocate memory to save all PFDI function status for all PE's */
+  g_pfdi_feature_check_details = (pfdi_feature_check_details *)
+                    val_memory_calloc(num_pe, sizeof(pfdi_feature_check_details));
+  if (g_pfdi_feature_check_details == NULL) {
     val_print(ACS_PRINT_ERR,
-              "\n       PFDI query PE Test Part function failed err = %d", test_count);
+                "\n       Allocation for PFDI Feature Check Function Failed", 0);
     val_set_status(index, RESULT_FAIL(TEST_NUM, 1));
     return;
   }
 
-  test_buffer->test_count = test_count;
-  val_data_cache_ops_by_va((addr_t)&test_buffer->test_count, CLEAN_AND_INVALIDATE);
+  for (i = 0; i < num_pe; i++) {
+    status_buffer = g_pfdi_feature_check_details + i;
+    val_data_cache_ops_by_va((addr_t)&status_buffer->status, CLEAN_AND_INVALIDATE);
+    val_data_cache_ops_by_va((addr_t)&status_buffer->status_invalid, CLEAN_AND_INVALIDATE);
+  }
 
-  /* Execute pfdi_version_check function in All PE's */
+  /* Invoke PFDI Feature function for current PE index */
+  check_feature();
+
+  /* Execute check_feature function in All PE's */
   for (i = 0; i < num_pe; i++) {
     if (i != index) {
       timeout = TIMEOUT_LARGE;
-      val_execute_on_pe(i, pfdi_test_part_count, 0);
+      val_execute_on_pe(i, check_feature, 0);
 
       while ((--timeout) && (IS_RESULT_PENDING(val_get_status(i))));
 
       if (timeout == 0) {
         val_print(ACS_PRINT_ERR, "\n       **Timed out** for PE index = %d", i);
         val_set_status(i, RESULT_FAIL(TEST_NUM, 2));
-        return;
+        goto free_pfdi_details;
       }
     }
   }
@@ -86,27 +103,34 @@ static void payload(uint32_t num_pe)
 
   /* Check return status of function for all PE's */
   for (i = 0; i < num_pe; i++) {
-    test_buffer = g_pfdi_pe_test_support_info + i;
+    status_buffer = g_pfdi_feature_check_details + i;
+    val_data_cache_ops_by_va((addr_t)&status_buffer->status, CLEAN_AND_INVALIDATE);
+    val_data_cache_ops_by_va((addr_t)&status_buffer->status_invalid, CLEAN_AND_INVALIDATE);
+    run_fail = 0;
 
-    val_data_cache_ops_by_va((addr_t)&test_buffer->test_count, CLEAN_AND_INVALIDATE);
-
-    if (test_buffer->test_count == PFDI_ACS_NOT_SUPPORTED) {
-      val_print(ACS_PRINT_ERR, "\n       PFDI query Test part count not supported", 0);
-      val_print(ACS_PRINT_ERR, "\n       Failed on PE = %d", i);
-      val_set_status(i, RESULT_FAIL(TEST_NUM, 3));
-    } else if (test_buffer->test_count >= PFDI_ACS_SUCCESS) {
-      val_print(ACS_PRINT_DEBUG, "\n       PE Index = %d", i);
-      val_print(ACS_PRINT_DEBUG,
-                    "\n       Test Part Count = 0x%llx", test_buffer->test_count);
-    } else {
-      val_print(ACS_PRINT_ERR, "\n       PFDI query Test part count failed err = %lld",
-                                  test_buffer->test_count);
-      val_print(ACS_PRINT_ERR, "\n       Failed on PE = %d", i);
-      val_set_status(i, RESULT_FAIL(TEST_NUM, 4));
+    if (status_buffer->status < PFDI_ACS_SUCCESS) {
+      val_print(ACS_PRINT_ERR, "\n       PFDI Feature Check function failed err = %ld",
+                                                        status_buffer->status);
+      val_print(ACS_PRINT_ERR, " on PE index = %d", i);
+      run_fail++;
     }
+
+    if (status_buffer->status_invalid != PFDI_ACS_NOT_SUPPORTED) {
+      val_print(ACS_PRINT_ERR, "\n       PFDI Feature Check invalid function failed err = %d",
+                                                        status_buffer->status_invalid);
+      val_print(ACS_PRINT_ERR, " on PE index = %d", i);
+      run_fail++;
+    }
+
+    if (run_fail)
+      val_set_status(i, RESULT_FAIL(TEST_NUM, 3));
+    else
+      val_set_status(i, RESULT_PASS(TEST_NUM, 1));
   }
 
-  val_set_status(index, RESULT_PASS(TEST_NUM, 1));
+free_pfdi_details:
+  val_memory_free((void *) g_pfdi_feature_check_details);
+
   return;
 }
 
@@ -117,7 +141,7 @@ uint32_t pfdi004_entry(uint32_t num_pe)
   status = val_initialize_test(TEST_NUM, TEST_DESC, num_pe);
 
   if (status != ACS_STATUS_SKIP)
-    payload(num_pe);
+    val_run_test_configurable_payload(&num_pe, payload_feature_check);
 
   /* get the result from all PE and check for failure */
   status = val_check_for_error(TEST_NUM, num_pe, TEST_RULE);
