+++ test_pool/pfdi/pfdi006.c
@@ -20,98 +20,63 @@
 #include "val/include/acs_memory.h"
 
 #define TEST_NUM   (ACS_PFDI_TEST_NUM_BASE + 6)
-#define TEST_RULE  "R0082"
-#define TEST_DESC  "Query PE boot test status                 "
+#define TEST_RULE  "R0071"
+#define TEST_DESC  "Check num of Test Part supported          "
 
-typedef struct{
-  int64_t results_status;
-  int64_t fault_id;
-} pfdi_results_status_details;
+typedef struct {
+  int64_t test_count;
+} pfdi_pe_test_support_info;
 
-pfdi_results_status_details *g_pfdi_results_status_details;
+pfdi_pe_test_support_info *g_pfdi_pe_test_support_info;
 
 void
-pfdi_test_results(void)
+pfdi_test_part_count(void)
 {
   uint32_t index = val_pe_get_index_mpid(val_pe_get_mpid());
-  int64_t  status, fault_test_id;
-  pfdi_results_status_details   *status_buffer;
+  pfdi_pe_test_support_info *test_buffer = g_pfdi_pe_test_support_info + index;
 
-  status_buffer = g_pfdi_results_status_details + index;
-
-  /* Invoke PFDI Results function for current PE index */
-  status = val_pfdi_pe_test_result(&fault_test_id);
-
-  status_buffer->results_status = status;
-  val_data_cache_ops_by_va((addr_t)&status_buffer->results_status, CLEAN_AND_INVALIDATE);
-  if (status < PFDI_ACS_SUCCESS) {
-    status_buffer->fault_id = fault_test_id;
-    val_data_cache_ops_by_va((addr_t)&status_buffer->fault_id, CLEAN_AND_INVALIDATE);
-    val_set_status(index, RESULT_FAIL(TEST_NUM, 1));
-    return;
-  }
+  /* Invoke PFDI Feature function for current PE index */
+  test_buffer->test_count = val_pfdi_pe_test_part_count();
+  val_data_cache_ops_by_va((addr_t)&test_buffer->test_count, CLEAN_AND_INVALIDATE);
 
   val_set_status(index, RESULT_PASS(TEST_NUM, 1));
   return;
 }
 
-static void payload_test_results(void *arg)
+static void payload_pe_test_info_check(void *arg)
 {
   uint32_t index = val_pe_get_index_mpid(val_pe_get_mpid());
+  uint32_t num_pe = *((uint32_t *)arg);
   uint32_t timeout, i = 0;
-  int64_t  status, fault_test_id;
-  pfdi_results_status_details *status_buffer;
-  uint32_t num_pe = *(uint32_t *)arg;
+  pfdi_pe_test_support_info *test_buffer;
 
-  /* Invoke PFDI Run count function for current PE index */
-  status = val_pfdi_pe_test_result(&fault_test_id);
-  if (status < PFDI_ACS_SUCCESS) {
-    if (status == PFDI_ACS_FAULT_FOUND) {
-      if (fault_test_id == PFDI_ACS_UNKNOWN) {
-        val_print(ACS_PRINT_ERR,
-                "\n       PFDI test part that triggered the fault cannot be identified", 0);
-      } else {
-        val_print(ACS_PRINT_ERR,
-                "\n       PFDI test part that triggered the fault = %d", fault_test_id);
-      }
-    } else if (status == PFDI_ACS_ERROR) {
-      val_print(ACS_PRINT_ERR,
-                "\n       PFDI Test part(s) have executed but failed to complete", 0);
-    } else {
-      val_print(ACS_PRINT_ERR,
-               "\n       PFDI query PE Results function failed err = %d", status);
-    }
-    val_set_status(index, RESULT_FAIL(TEST_NUM, 1));
-    return;
-  }
-
-  /* Allocate memory to save all PFDI results status and fault id's for all PE's */
-  g_pfdi_results_status_details = (pfdi_results_status_details *)
-                    val_memory_calloc(num_pe, sizeof(pfdi_results_status_details));
-  if (g_pfdi_results_status_details == NULL) {
-    val_print(ACS_PRINT_ERR,
-                "\n       Allocation for PFDI Results Function Failed", 0);
-    val_set_status(index, RESULT_FAIL(TEST_NUM, 2));
-    return;
+  /* Allocate memory to save all PFDI Versions or status for all PE's */
+  g_pfdi_pe_test_support_info =
+        (pfdi_pe_test_support_info *)val_memory_calloc(num_pe, sizeof(pfdi_pe_test_support_info));
+  if (g_pfdi_pe_test_support_info == NULL) {
+      val_print(ACS_PRINT_ERR, "\n       Allocation for PFDI PE Test Support Info Failed \n", 0);
+      return;
   }
 
   for (i = 0; i < num_pe; i++) {
-    status_buffer = g_pfdi_results_status_details + i;
-    val_data_cache_ops_by_va((addr_t)&status_buffer->results_status, CLEAN_AND_INVALIDATE);
-    val_data_cache_ops_by_va((addr_t)&status_buffer->fault_id, CLEAN_AND_INVALIDATE);
+    test_buffer = g_pfdi_pe_test_support_info + i;
+    val_data_cache_ops_by_va((addr_t)&test_buffer->test_count, CLEAN_AND_INVALIDATE);
   }
 
-  /* Execute pfdi_test_results function in All PE's */
+  /* Invoke PFDI Test Part count function for current PE index */
+  pfdi_test_part_count();
+
+  /* Execute pfdi_version_check function in All PE's */
   for (i = 0; i < num_pe; i++) {
     if (i != index) {
       timeout = TIMEOUT_LARGE;
-      val_execute_on_pe(i, pfdi_test_results, 0);
+      val_execute_on_pe(i, pfdi_test_part_count, 0);
 
       while ((--timeout) && (IS_RESULT_PENDING(val_get_status(i))));
 
       if (timeout == 0) {
         val_print(ACS_PRINT_ERR, "\n       **Timed out** for PE index = %d", i);
-        val_set_status(i, RESULT_FAIL(TEST_NUM, 3));
+        val_set_status(i, RESULT_FAIL(TEST_NUM, 2));
         goto free_pfdi_details;
       }
     }
@@ -120,38 +85,26 @@ static void payload_test_results(void *arg)
 
   /* Check return status of function for all PE's */
   for (i = 0; i < num_pe; i++) {
-    if (i != index) {
-      status_buffer = g_pfdi_results_status_details + i;
-      val_data_cache_ops_by_va((addr_t)&status_buffer->results_status, CLEAN_AND_INVALIDATE);
-      val_data_cache_ops_by_va((addr_t)&status_buffer->fault_id, CLEAN_AND_INVALIDATE);
+    test_buffer = g_pfdi_pe_test_support_info + i;
 
-      if (status_buffer->results_status < PFDI_ACS_SUCCESS) {
-        if (status_buffer->results_status == PFDI_ACS_FAULT_FOUND) {
-          if (status_buffer->fault_id == PFDI_ACS_UNKNOWN) {
-            val_print(ACS_PRINT_ERR,
-                "\n       PFDI test part that triggered the fault cannot be identified", 0);
-          } else {
-            val_print(ACS_PRINT_ERR,
-                "\n       PFDI test part that triggered the fault = %d", status_buffer->fault_id);
-          }
-        } else if (status_buffer->results_status == PFDI_ACS_ERROR) {
-          val_print(ACS_PRINT_ERR,
-                "\n       PFDI Test part(s) have executed but failed to complete", 0);
-        } else {
-          val_print(ACS_PRINT_ERR, "\n       PFDI query PE Results function failed err = %d",
-                                                        status_buffer->results_status);
-        }
-        val_set_status(i, RESULT_FAIL(TEST_NUM, 4));
-        goto free_pfdi_details;
-      }
+    val_data_cache_ops_by_va((addr_t)&test_buffer->test_count, CLEAN_AND_INVALIDATE);
 
+    if (test_buffer->test_count < PFDI_ACS_SUCCESS) {
+      val_print(ACS_PRINT_ERR, "\n       PFDI query Test part count failed %lld",
+                                  test_buffer->test_count);
+      val_print(ACS_PRINT_ERR, "on PE = %d", i);
+      val_set_status(i, RESULT_FAIL(TEST_NUM, 3));
+    } else {
+      val_print(ACS_PRINT_DEBUG,
+                    "\n       Test Part Count = 0x%llx", test_buffer->test_count);
+      val_print(ACS_PRINT_DEBUG, "on PE  %d", i);
+      val_set_status(i, RESULT_PASS(TEST_NUM, 1));
     }
   }
 
-  val_set_status(index, RESULT_PASS(TEST_NUM, 1));
-
 free_pfdi_details:
-  val_memory_free((void *) g_pfdi_results_status_details);
+  val_memory_free((void *) g_pfdi_pe_test_support_info);
+
   return;
 }
 
@@ -162,7 +115,7 @@ uint32_t pfdi006_entry(uint32_t num_pe)
   status = val_initialize_test(TEST_NUM, TEST_DESC, num_pe);
 
   if (status != ACS_STATUS_SKIP)
-    val_run_test_configurable_payload(&num_pe, payload_test_results);
+    val_run_test_configurable_payload(&num_pe, payload_pe_test_info_check);
 
   /* get the result from all PE and check for failure */
   status = val_check_for_error(TEST_NUM, num_pe, TEST_RULE);
