+++ test_pool/pfdi/pfdi008.c
@@ -20,91 +20,74 @@
 #include "val/include/acs_memory.h"
 
 #define TEST_NUM   (ACS_PFDI_TEST_NUM_BASE + 8)
-#define TEST_RULE  "R0060"
-#define TEST_DESC  "Check PFDI Discovery Function Support     "
+#define TEST_RULE  "R0082"
+#define TEST_DESC  "Query PE boot test status                 "
 
 typedef struct{
-  int64_t status;
-  int64_t status_invalid;
-} pfdi_feature_check_details;
+  int64_t results_status;
+  int64_t fault_id;
+} pfdi_results_status_details;
 
-pfdi_feature_check_details *g_pfdi_feature_check_details;
-
+pfdi_results_status_details *g_pfdi_results_status_details;
 
 void
-check_feature()
+pfdi_test_results(void)
 {
   uint32_t index = val_pe_get_index_mpid(val_pe_get_mpid());
-  int64_t  status;
-  pfdi_feature_check_details  *status_buffer;
+  int64_t fault_test_id;
+  pfdi_results_status_details   *status_buffer;
 
-  status_buffer = g_pfdi_feature_check_details + index;
+  status_buffer = g_pfdi_results_status_details + index;
 
-  /* Invoke PFDI Feature function for current PE index */
-  status = val_pfdi_features(PFDI_FN_PFDI_FEATURES);
+  /* Invoke PFDI Results function for current PE index */
+  status_buffer->results_status = val_pfdi_pe_test_result(&fault_test_id);
 
-  /*Save the return status*/
-  status_buffer->status = status;
-  val_data_cache_ops_by_va((addr_t)&status_buffer->status, CLEAN_AND_INVALIDATE);
-
-  if (status != PFDI_ACS_SUCCESS) {
-    val_set_status(index, RESULT_FAIL(TEST_NUM, 1));
-    return;
-  }
-  /* Invoke PFDI Feature function with invalid id for current PE index */
-  status = val_pfdi_features(PFDI_FN_PFDI_INVALID);
-
-  /*Save the return status*/
-  status_buffer->status_invalid = status;
-  val_data_cache_ops_by_va((addr_t)&status_buffer->status_invalid, CLEAN_AND_INVALIDATE);
-
-  if (status != PFDI_ACS_NOT_SUPPORTED) {
-    val_set_status(index, RESULT_FAIL(TEST_NUM, 2));
-    return;
-  }
+  status_buffer->fault_id = fault_test_id;
+  val_data_cache_ops_by_va((addr_t)&status_buffer->results_status, CLEAN_AND_INVALIDATE);
+  val_data_cache_ops_by_va((addr_t)&status_buffer->fault_id, CLEAN_AND_INVALIDATE);
 
   val_set_status(index, RESULT_PASS(TEST_NUM, 1));
-
   return;
 }
 
-static void payload_feature_check(void *arg)
+static void payload_test_results(void *arg)
 {
   uint32_t index = val_pe_get_index_mpid(val_pe_get_mpid());
-  uint32_t timeout, i = 0, run_fail = 0;
-  pfdi_feature_check_details *status_buffer;
+  uint32_t timeout, i = 0, test_fail = 0;
+  pfdi_results_status_details *status_buffer;
   uint32_t num_pe = *(uint32_t *)arg;
 
-  /* Allocate memory to save all PFDI function status for all PE's */
-  g_pfdi_feature_check_details = (pfdi_feature_check_details *)
-                    val_memory_calloc(num_pe, sizeof(pfdi_feature_check_details));
-  if (g_pfdi_feature_check_details == NULL) {
+  /* Allocate memory to save all PFDI results status and fault id's for all PE's */
+  g_pfdi_results_status_details = (pfdi_results_status_details *)
+                    val_memory_calloc(num_pe, sizeof(pfdi_results_status_details));
+  if (g_pfdi_results_status_details == NULL) {
     val_print(ACS_PRINT_ERR,
-                "\n       Allocation for PFDI Feature Check Function Failed", 0);
+                "\n       Allocation for PFDI Results Function Failed", 0);
     val_set_status(index, RESULT_FAIL(TEST_NUM, 1));
     return;
   }
 
   for (i = 0; i < num_pe; i++) {
-    status_buffer = g_pfdi_feature_check_details + i;
-    val_data_cache_ops_by_va((addr_t)&status_buffer->status, CLEAN_AND_INVALIDATE);
-    val_data_cache_ops_by_va((addr_t)&status_buffer->status_invalid, CLEAN_AND_INVALIDATE);
+    status_buffer = g_pfdi_results_status_details + i;
+    val_data_cache_ops_by_va((addr_t)&status_buffer->results_status, CLEAN_AND_INVALIDATE);
+    val_data_cache_ops_by_va((addr_t)&status_buffer->fault_id, CLEAN_AND_INVALIDATE);
   }
 
-  /* Invoke PFDI Feature function for current PE index */
-  check_feature();
+  /* Invoke PFDI Run function for current PE index */
+  pfdi_test_results();
 
-  /* Execute check_feature function in All PE's */
+  /* Execute pfdi_test_results function in All PE's */
   for (i = 0; i < num_pe; i++) {
     if (i != index) {
       timeout = TIMEOUT_LARGE;
-      val_execute_on_pe(i, check_feature, 0);
+      val_execute_on_pe(i, pfdi_test_results, 0);
 
       while ((--timeout) && (IS_RESULT_PENDING(val_get_status(i))));
 
       if (timeout == 0) {
         val_print(ACS_PRINT_ERR, "\n       **Timed out** for PE index = %d", i);
         val_set_status(i, RESULT_FAIL(TEST_NUM, 2));
+        goto free_pfdi_details;
       }
     }
   }
@@ -112,32 +95,39 @@ static void payload_feature_check(void *arg)
 
   /* Check return status of function for all PE's */
   for (i = 0; i < num_pe; i++) {
-    status_buffer = g_pfdi_feature_check_details + i;
-    val_data_cache_ops_by_va((addr_t)&status_buffer->status, CLEAN_AND_INVALIDATE);
-    val_data_cache_ops_by_va((addr_t)&status_buffer->status_invalid, CLEAN_AND_INVALIDATE);
+    status_buffer = g_pfdi_results_status_details + i;
+    val_data_cache_ops_by_va((addr_t)&status_buffer->results_status, CLEAN_AND_INVALIDATE);
+    val_data_cache_ops_by_va((addr_t)&status_buffer->fault_id, CLEAN_AND_INVALIDATE);
+    test_fail = 0;
 
-    if (status_buffer->status < PFDI_ACS_SUCCESS) {
-      val_print(ACS_PRINT_ERR, "\n       PFDI query Feature Check function failed err = %d",
-                                                        status_buffer->status);
-      val_print(ACS_PRINT_ERR, " on PE index = %d", i);
+    if (status_buffer->results_status < PFDI_ACS_SUCCESS) {
+      if (status_buffer->results_status == PFDI_ACS_FAULT_FOUND) {
+        if (status_buffer->fault_id == PFDI_ACS_UNKNOWN) {
+          val_print(ACS_PRINT_ERR, "\n       Fault in PFDI test part on PE %d ", i);
+          val_print(ACS_PRINT_ERR, "cannot be identified", 0);
+        } else {
+          val_print(ACS_PRINT_ERR, "\n       PFDI test part %d ", status_buffer->fault_id);
+          val_print(ACS_PRINT_ERR, "triggered the fault on PE %d", i);
+        }
+      } else if (status_buffer->results_status == PFDI_ACS_ERROR) {
+        val_print(ACS_PRINT_ERR,
+              "\n       PFDI Test parts have executed but failed to complete on PE %d", i);
+      } else {
+        val_print(ACS_PRINT_ERR, "\n       PFDI PE Results function failed err = %d",
+                                                      status_buffer->results_status);
+        val_print(ACS_PRINT_ERR, "on PE  %d", i);
+      }
+      test_fail++;
+    }
+
+    if (test_fail)
       val_set_status(i, RESULT_FAIL(TEST_NUM, 3));
-      run_fail++;
-    }
-
-    if (status_buffer->status_invalid != PFDI_ACS_NOT_SUPPORTED) {
-      val_print(ACS_PRINT_ERR, "\n       PFDI query Feature Check invalid function failed err = %d",
-                                                        status_buffer->status_invalid);
-      val_print(ACS_PRINT_ERR, " on PE index = %d", i);
-      val_set_status(i, RESULT_FAIL(TEST_NUM, 4));
-      run_fail++;
-    }
+    else
+      val_set_status(i, RESULT_PASS(TEST_NUM, 1));
   }
 
-  if (run_fail == 0)
-    val_set_status(index, RESULT_PASS(TEST_NUM, 1));
-
-  val_memory_free((void *) g_pfdi_feature_check_details);
-
+free_pfdi_details:
+  val_memory_free((void *) g_pfdi_results_status_details);
   return;
 }
 
@@ -148,7 +138,7 @@ uint32_t pfdi008_entry(uint32_t num_pe)
   status = val_initialize_test(TEST_NUM, TEST_DESC, num_pe);
 
   if (status != ACS_STATUS_SKIP)
-    val_run_test_configurable_payload(&num_pe, payload_feature_check);
+    val_run_test_configurable_payload(&num_pe, payload_test_results);
 
   /* get the result from all PE and check for failure */
   status = val_check_for_error(TEST_NUM, num_pe, TEST_RULE);
