+++ val/src/pfdi_execute_test.c
@@ -62,6 +62,7 @@ val_pfdi_execute_pfdi_tests(uint32_t num_pe)
   status  = pfdi006_entry(num_pe);
   status  = pfdi007_entry(num_pe);
   status  = pfdi008_entry(num_pe);
+  status  = pfdi009_entry(num_pe);
 
   val_print_test_end(status, "PFDI");
